+++ src/reducer.test.js
@@ -36,7 +36,6 @@ const questions = [
   },
 ];
 
-
 it(`Reducer without additional parameters should return initial state`, () => {
   expect(reducer(void 0, {})).toEqual({
     step: -1,
@@ -52,8 +51,7 @@ it(`Reducer should increment current step by a given value`, () => {
     mistakes: 0,
     questions,
   }, {
-    type: ActionType.INCREMENT_STEP,
-    payload: 1,
+    type: ActionType.INCREMENT_STEP
   })).toEqual({
     step: 0,
     mistakes: 0,
@@ -61,14 +59,14 @@ it(`Reducer should increment current step by a given value`, () => {
   });
 
   expect(reducer({
-    step: -1,
+    step: 0,
     mistakes: 0,
     questions,
   }, {
     type: ActionType.INCREMENT_STEP,
     payload: 0,
   })).toEqual({
-    step: -1,
+    step: 1,
     mistakes: 0,
     questions,
   });
@@ -101,114 +99,21 @@ it(`Reducer should increment number of mistakes by a given value`, () => {
 describe(`Action creators work correctly`, () => {
   it(`Action creator for incrementing step returns correct action`, () => {
     expect(ActionCreator.incrementStep()).toEqual({
-      type: ActionType.INCREMENT_STEP,
-      payload: 1,
-    });
-  });
-
-  it(`Action creator for incrementing mistake returns action with 0 payload if answer for artist is correct`, () => {
-    expect(ActionCreator.incrementMistake({
-      type: `artist`,
-      song: {
-        artist: `correct`,
-        src: ``,
-      },
-      answers: [
-        {
-          artist: `correct`,
-          picture: ``,
-        }, {
-          artist: `incorrect`,
-          picture: ``,
-        }, {
-          artist: `incorrect-2`,
-          picture: ``,
-        },
-      ]
-    }, {
-      artist: `correct`,
-      picture: ``,
-    })).toEqual({
-      type: ActionType.INCREMENT_MISTAKES,
-      payload: 0,
+      type: ActionType.INCREMENT_STEP
     });
   });
 
   it(`Action creator for incrementing mistake returns action with 1 payload if answer for artist is incorrect`, () => {
-    expect(ActionCreator.incrementMistake({
-      type: `artist`,
-      song: {
-        artist: `correct`,
-        src: ``,
-      },
-      answers: [
-        {
-          artist: `correct`,
-          picture: ``,
-        }, {
-          artist: `incorrect`,
-          picture: ``,
-        }, {
-          artist: `incorrect-2`,
-          picture: ``,
-        },
-      ]
-    }, {
-      artist: `incorrect`,
-      picture: ``,
-    })).toEqual({
+    expect(ActionCreator.incrementMistake(false)).toEqual({
       type: ActionType.INCREMENT_MISTAKES,
       payload: 1,
     });
   });
 
   it(`Action creator for incrementing mistake returns action with 0 payload if answer for genre is correct`, () => {
-    expect(ActionCreator.incrementMistake({
-      type: `genre`,
-      genre: `jazz`,
-      answers: [
-        {
-          genre: `rock`,
-          src: ``,
-        }, {
-          genre: `jazz`,
-          src: ``,
-        }, {
-          genre: `blues`,
-          src: ``,
-        }, {
-          genre: `blues`,
-          src: ``,
-        },
-      ]
-    }, [false, true, false, false])).toEqual({
+    expect(ActionCreator.incrementMistake(true)).toEqual({
       type: ActionType.INCREMENT_MISTAKES,
       payload: 0,
-    });
-  });
-
-  it(`Action creator for incrementing mistake returns action with 1 payload if answer for genre is incorrect`, () => {
-    expect(ActionCreator.incrementMistake({
-      type: `genre`,
-      genre: `jazz`,
-      answers: [
-        {
-          genre: `blues`,
-          src: ``,
-        }, {
-          genre: `blues`,
-          src: ``,
-        }, {
-          genre: `blues`,
-          src: ``,
-        }, {
-          genre: `blues`,
-          src: ``,
-        },
-      ]
-    }, [true, true, true, true])).toEqual({
-      type: ActionType.INCREMENT_MISTAKES,
-      payload: 1,
     });
   });
 });
